requirements.txt
python-telegram-bot>=21.0
yfinance
pandas
ta
python-dotenv
fastapi
uvicorn
sse-starlette
